refactor: jdbc member repository 예외 사항 추가
findById(), deleteById() 각각 예외 사항 추가
추가된 예외 사항에 대한 테스트 케이스 추가

+++ src/main/java/com/study/db_connection/repository/jdbc/JdbcMemberRepository.java
@@ -66,6 +66,9 @@ public class JdbcMemberRepository {
     }
 
     public Member findById(Long id) {
+        if (id == null) {
+            throw new IllegalStateException("id not nullable");
+        }
         String sql = "select * from member where id = ?";
 
         Connection con = null;
@@ -147,6 +150,9 @@ public class JdbcMemberRepository {
     }
 
     public void deleteById(Long id) {
+        if (id == null) {
+            throw new IllegalStateException("id not nullable");
+        }
         String sql = "delete from member where id = ?";
 
         Connection con = null;
